[JdeveloperAcc] UT in conform with: Framework-level dependency injection supports the following scopes: Singleton - always return the same instance. Transient - return new instance every time. Scoped - return same instance in current (request) scope.
DbContext is Scoped in .NET Core.

https://learn.microsoft.com/en-us/ef/core/dbcontext-configuration/

A typical unit-of-work when using Entity Framework Core (EF Core) involves:

Creation of a DbContext instance
Tracking of entity instances by the context. Entities become tracked by
Being returned from a query
Being added or attached to the context
Changes are made to the tracked entities as needed to implement the business rule
SaveChanges or SaveChangesAsync is called. EF Core detects the changes made and writes them to the database.
The DbContext instance is disposed